# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Mealswapp API
  version: 0.3.0
  description: Shared gateway source of truth for Phase 03 authentication and gateway contracts.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
security: []
paths:
...
  /api/v1/profile:
    get:
      operationId: getApiV1Profile
      summary: Read authenticated profile and preferences
      security:
        - cookieAuth: []
      responses:
        "200":
          $ref: "#/components/responses/Profile"
        "401":
          $ref: "#/components/responses/Error"
    put:
      operationId: putApiV1Profile
      summary: Update authenticated profile preferences
      security:
        - cookieAuth: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/Profile"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
...
  /api/v1/daily-diets:
    get:
      operationId: getApiV1DailyDiets
      summary: List the authenticated user's saved daily diets
      security:
        - cookieAuth: []
      responses:
        "200":
          $ref: "#/components/responses/DailyDietCollection"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
    post:
      operationId: postApiV1DailyDiet
      summary: Create an authenticated user's saved daily diet
      description: Creates a user-owned collection. Names are unique per user after trimming and case folding. The server derives ownership and aggregate macros from the authenticated session and persisted Food Object data; clients cannot submit user IDs or authoritative totals.
      security:
        - cookieAuth: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyDietCreateRequest"
      responses:
        "201":
          $ref: "#/components/responses/DailyDiet"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
  # Implements DESIGN-008 SavedDataRepository daily-diet item contract.
  /api/v1/daily-diets/{dietId}:
    parameters:
      - $ref: "#/components/parameters/DailyDietId"
    get:
      operationId: getApiV1DailyDiet
      summary: Read one authenticated user's saved daily diet
      security:
        - cookieAuth: []
      responses:
        "200":
          $ref: "#/components/responses/DailyDiet"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
    put:
      operationId: putApiV1DailyDiet
      summary: Replace an authenticated user's saved daily diet
      description: Replaces the user-owned collection and returns server-derived aggregate macros. Names are unique per user after trimming and case folding. The request contains Food Object quantities only; it cannot override ownership or totals.
      security:
        - cookieAuth: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyDietReplaceRequest"
      responses:
        "200":
          $ref: "#/components/responses/DailyDiet"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteApiV1DailyDiet
      summary: Delete an authenticated user's saved daily diet
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        "204":
          description: Saved daily diet deleted.
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
  # Implements DESIGN-008 ProfileController authenticated custom-item contract.
  /api/v1/custom-items:
    post:
      operationId: postApiV1CustomItem
      summary: Create an authenticated user's private custom food item
      description: Ownership is derived only from the authenticated session. Exact retries with the same Idempotency-Key and normalized body replay the original response; changed normalized bodies return 409.
      security:
        - cookieAuth: []
          csrfHeader: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomItemRequest"
      responses:
        "201":
          $ref: "#/components/responses/CustomItem"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
  # Implements DESIGN-008 ProfileController owner-scoped custom-item contract.
  /api/v1/custom-items/{itemId}:
    parameters:
      - $ref: "#/components/parameters/CustomItemId"
    get:
      operationId: getApiV1CustomItem
      summary: Read one authenticated user's private custom food item
      security:
        - cookieAuth: []
      responses:
        "200":
          $ref: "#/components/responses/CustomItem"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
    put:
      operationId: putApiV1CustomItem
      summary: Replace one authenticated user's private custom food item
      security:
        - cookieAuth: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomItemRequest"
      responses:
        "200":
          $ref: "#/components/responses/CustomItem"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteApiV1CustomItem
      summary: Delete one authenticated user's private custom food item
      security:
        - cookieAuth: []
          csrfHeader: []
      responses:
        "204":
          description: Private custom food item deleted.
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
        "504":
          $ref: "#/components/responses/Error"
  # Implements DESIGN-004 JobStatusTracker optimization submission contract.
...
components:
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable client-generated key for retrying an authenticated mutation. The key is scoped to the authenticated user, HTTP method, route, and normalized request body; it never carries a user ID.
      schema:
        type: string
        minLength: 8
        maxLength: 255
...
    DailyDietId:
      name: dietId
      in: path
      required: true
      description: User-owned saved daily-diet identifier. Ownership is derived from the authenticated session.
      schema:
        type: string
        format: uuid
    CustomItemId:
      name: itemId
      in: path
      required: true
      description: Private custom-item identifier. Ownership is derived from the authenticated session and cross-user IDs return not found.
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mealswapp_access
    refreshCookie:
      type: apiKey
      in: cookie
      name: mealswapp_refresh
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
  responses:
...
    Error:
      description: User-safe gateway error envelope with no client-visible audit snapshot, raw provider payload, or internal diagnostics.
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    # Implements DESIGN-004 JobStatusTracker and DESIGN-010 RateLimiter shared admission contract.
...
    Profile:
      description: Profile envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProfileEnvelope"
...
    DailyDiet:
      description: User-owned saved daily-diet collection with server-derived aggregate macros.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DailyDietEnvelope"
    DailyDietCollection:
      description: User-owned saved daily-diet collection list.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DailyDietCollectionEnvelope"
    CustomItem:
      description: Authenticated user's private custom food item without ownership metadata.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/CustomItemEnvelope"
...
  schemas:
    AppError:
      type: object
      additionalProperties: false
      required: [category, code, message, retryable]
      properties:
        category:
          type: string
          enum: [validation, auth, entitlement, security, network, timeout, server, dependency, rate_limit, unknown]
        code:
          type: string
          minLength: 1
          maxLength: 120
        message:
          type: string
          minLength: 1
          maxLength: 500
        retryable:
          type: boolean
        requestId:
          type: string
          minLength: 1
          maxLength: 120
    # Implements DESIGN-009 AdminController audit-safe error boundary.
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required: [status, requestId, error]
      properties:
        status:
          type: string
          const: error
        requestId:
          type: string
          maxLength: 120
        error:
          $ref: "#/components/schemas/AppError"
...
    ProfileUpdateRequest:
      type: object
      additionalProperties: false
      required: [unitSystem, themePreference]
      properties:
        displayName:
          type: string
        unitSystem:
          type: string
          enum: [metric, imperial]
        themePreference:
          type: string
          enum: [system, light, dark]
    ProfileEnvelope:
      type: object
      required: [status, requestId, data]
      properties:
        status:
          type: string
          enum: [ok]
        requestId:
          type: string
        data:
          type: object
          required: [userId, displayName, unitSystem, themePreference, requiresUnitRecalculation]
          properties:
            userId:
              type: string
              format: uuid
            displayName:
              type: string
            unitSystem:
              type: string
              enum: [metric, imperial]
            themePreference:
              type: string
              enum: [system, light, dark]
            requiresUnitRecalculation:
              type: boolean
...
    CanonicalQuantityUnit:
      type: string
      enum: [g, ml, oz, fl_oz]
      description: Canonical mass or volume unit accepted by saved-diet entries.
    FoodObjectType:
      type: string
      enum: [food_item, meal]
    FoodObjectQuantity:
      type: object
      additionalProperties: false
      required: [foodObjectId, foodObjectType, quantity, unit, position]
      properties:
        foodObjectId:
          type: string
          format: uuid
        foodObjectType:
          $ref: "#/components/schemas/FoodObjectType"
        quantity:
          type: number
          exclusiveMinimum: 0
          maximum: 1000000
          multipleOf: 0.001
        unit:
          $ref: "#/components/schemas/CanonicalQuantityUnit"
        position:
          type: integer
          minimum: 0
          maximum: 99
...
    DailyDietFoodObjectEntry:
      type: object
      additionalProperties: false
      required: [id, foodObjectId, foodObjectType, quantity, unit, position]
      properties:
        id:
          type: string
          format: uuid
        foodObjectId:
          type: string
          format: uuid
        foodObjectType:
          $ref: "#/components/schemas/FoodObjectType"
        quantity:
          type: number
          exclusiveMinimum: 0
          maximum: 1000000
          multipleOf: 0.001
        unit:
          $ref: "#/components/schemas/CanonicalQuantityUnit"
        position:
          type: integer
          minimum: 0
          maximum: 99
    MacroProjection:
      type: object
      additionalProperties: false
      required: [protein, carbohydrates, fat, calories]
      description: Server-derived one-day aggregate macro and calorie projection; clients cannot submit or override these values.
      properties:
        protein:
          type: number
          minimum: 0
          maximum: 1000000000
        carbohydrates:
          type: number
          minimum: 0
          maximum: 1000000000
        fat:
          type: number
          minimum: 0
          maximum: 1000000000
        calories:
          type: number
          minimum: 0
          maximum: 1000000000
    DailyDiet:
      type: object
      additionalProperties: false
      required: [id, name, entries, aggregateMacros, createdAt, updatedAt]
      description: User-owned saved daily diet. The authenticated session, not a client field, determines ownership.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: User-scoped unique name, compared after trimming and case folding.
        entries:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/DailyDietFoodObjectEntry"
        aggregateMacros:
          $ref: "#/components/schemas/MacroProjection"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DailyDietCreateRequest:
      type: object
      additionalProperties: false
      required: [name, entries]
      description: Client-editable saved-diet fields. It intentionally has no userId or aggregate macro fields.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: User-scoped unique name, compared after trimming and case folding.
        entries:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/FoodObjectQuantity"
    DailyDietReplaceRequest:
      type: object
      additionalProperties: false
      required: [name, entries]
      description: Client-editable replacement fields. Aggregate macros are recalculated by the server.
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
          description: User-scoped unique name, compared after trimming and case folding.
        entries:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/FoodObjectQuantity"
    DailyDietEnvelope:
      type: object
      required: [status, requestId, data]
      properties:
        status:
          type: string
          enum: [ok]
        requestId:
          type: string
        data:
          $ref: "#/components/schemas/DailyDiet"
    DailyDietCollectionEnvelope:
      type: object
      required: [status, requestId, data]
      properties:
        status:
          type: string
          enum: [ok]
        requestId:
          type: string
        data:
          type: object
          additionalProperties: false
          required: [diets]
          properties:
            diets:
              type: array
              items:
                $ref: "#/components/schemas/DailyDiet"
    # Implements DESIGN-008 ProfileController shared custom-item fields.
    CustomItemFields:
      type: object
      required: [name, physicalState, macrosPer100, micros]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          pattern: '^[^\x00]*[^\s\x00][^\x00]*$'
        physicalState:
          type: string
          enum: [solid, liquid]
        prepTimeMinutes:
          type: integer
          minimum: 0
        averageUnitWeightGrams:
          type: number
          exclusiveMinimum: 0
        averageServingVolumeMilliliters:
          type: number
          exclusiveMinimum: 0
        densityGramsPerMilliliter:
          type: number
          exclusiveMinimum: 0
        densitySourceProvider:
          type: string
          maxLength: 200
          pattern: '^[^\x00]*$'
        densitySourceFoodId:
          type: string
          maxLength: 200
          pattern: '^[^\x00]*$'
        densitySourceKind:
          type: string
          enum: [imported, manual, estimated]
          pattern: '^[^\x00]*$'
        macrosPer100:
          $ref: "#/components/schemas/MacroProfile"
        micros:
          type: object
          propertyNames:
            minLength: 1
            maxLength: 120
            pattern: '^[^\x00]*$'
          additionalProperties:
            type: number
            minimum: 0
        foodCategoryIds:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
        culinaryRoleIds:
          type: array
          maxItems: 100
          uniqueItems: true
          items:
            type: string
            format: uuid
        imageUrl:
          type: string
          format: uri-reference
          maxLength: 2048
          pattern: '^[^\x00]*$'
    # Implements DESIGN-008 ProfileController client-editable custom-item contract.
    CustomItemRequest:
      allOf:
        - $ref: "#/components/schemas/CustomItemFields"
      unevaluatedProperties: false
      description: Client-editable private item fields. userId and ownerId are intentionally unsupported.
    # Implements DESIGN-008 ProfileController and DataExporter private-item projection.
    CustomItem:
      allOf:
        - $ref: "#/components/schemas/CustomItemFields"
        - type: object
          required: [id, prepTimeMinutes, micros, foodCategories, culinaryRoles]
          properties:
            id:
              type: string
              format: uuid
            foodCategories:
              type: array
              items:
                $ref: "#/components/schemas/ClassificationSummary"
            culinaryRoles:
              type: array
              items:
                $ref: "#/components/schemas/ClassificationSummary"
      unevaluatedProperties: false
    CustomItemEnvelope:
      type: object
      additionalProperties: false
      required: [status, requestId, data]
      properties:
        status:
          type: string
          const: ok
        requestId:
          type: string
        data:
          $ref: "#/components/schemas/CustomItem"
    # Implements DESIGN-004 JobStatusTracker optimization contract.
...
    ClassificationSummary:
      type: object
      additionalProperties: false
      required: [id, name, kind]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
        kind:
          type: string
          enum: [food_category, culinary_role]
    # Implements DESIGN-002 SearchController macro profile contract.
    MacroProfile:
      type: object
      additionalProperties: false
      required: [protein, carbohydrates, fat]
      properties:
        protein:
          type: number
          minimum: 0
        carbohydrates:
          type: number
          minimum: 0
        fat:
          type: number
          minimum: 0
    # Implements DESIGN-002 SearchController substitution source summary contract.
...